Make right DB and rquery optional (#23)
* Make _right and _rquery optional

There are times when a user may only want to run checks against a
single source database. In this case, requiring that a second database
be included is unnecessary and would require either duplication of work
or some sort of complicated workaround.

These changes make the `right` database (and subsequently `rquery`)
optional when instantiating a QueryPair, Comparator, and ComparatorSet.
The order of arguments has also changed for QueryPair, moving `right`
after `lquery` to allow for a `None` default. The class can now be
instantiated in this way: `QueryPair(left, lquery, right, rquery).

* Add CHANGES and version bump