Add better transaction handling to Db objects
Transactions on PostgresDb connections were never closed under the
previous configuration unless close() was called explicitly by the
end user. This could lead to long-running connections with negative
effects on DB performance or on things like read-replica lag.

These changes add an implicit close() to each call to the query() and
execute() methods, as well as a transaction block on execute(). There
is also some general cleanup on the way connect() is called throughout
the Db classes.