Gracefully handle QueryResultCol rquery format (#27)
Sometimes mixed or empty results could be returned by the lquery when
trying to format the rquery as a part of the QueryPair. These changes
separate the __str__ method of QueryResultCol from a new method that
can be used by the QuerPair to get a valid string to use to format
the rquery. The new _rquery_format method filters out None vaues and
properly returns a single-value result without a trailing tuple comma.